feature: refactor default layout
